Move out numeric methods (sum, mean)
